+++ yAppWeb/ReleaseNotes.md
@@ -10,3 +10,4 @@
 * devTaskIssue#45-FeatureIssue#1-Setup a post table schema using C#
 * devTaskIssue#70-FeatureIssue#5-Add unit tests for all friendship actions
 * devTaskIssue#73-FeatureIssue#5-Add unit tests for all cognito actions
+* devTaskIssue#74-FeatureIssue#5-Add integration tests for all cognito actions